+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-## Unreleased
+## 5.10.0 (2023-01-08)
 
 #### Features
 
-- add Intellijj Http Graphql Method support AnWeber/vscode-httpyac#165
+- add Intellijj Http Graphql Method support (AnWeber/vscode-httpyac#165)
 - update Intellijj Dynamic Variables
 - add Intellij PreRequest Script Support
 - add `$random` Utils to create Mock Data